pyproject.toml
[project]
name = "gork-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Haris Mehuljic", email = "haris.mehuljic@drake.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "google-api-python-client>=2.176.0",
    "openai>=1.95.1",
    "pillow>=11.3.0",
]

[project.scripts]
start = "gork_bot.app:main"
testing = "gork_bot.app:testing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["dotenv"]
